+++ bridge/opentracing/internal/mock.go
@@ -26,6 +26,7 @@ import (
 	otelparent "go.opentelemetry.io/otel/internal/trace/parent"
 	"go.opentelemetry.io/otel/label"
 	"go.opentelemetry.io/otel/trace"
+	"go.opentelemetry.io/otel/trace/codes"
 
 	"go.opentelemetry.io/otel/bridge/opentracing/migration"
 )