+++ CHANGELOG.md
@@ -11,6 +11,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - Reverse order in which `Resource` attributes are merged, per change in spec. (#1501)
+- Make `labels` into a module.
+- Move `semconv.SpanStatusFromHTTPStatusCode` into `codes` and make `semconv` into a module.
+- Move `codes` into `trace/codes` and make `trace` into a module.
 - Rename `internal/baggage` to `baggage/updatable`.
 
 ## [0.16.0] - 2020-01-13